Update README.md
i was wrong about this. actually even if i could steal an ithing, i can't try this out since the engine isn't available in the app store. you have to compile it yourself. you need to buy le expensive macbook because xcode requires it in order to compile the engine! i love the forced consumerism! i love apple as a company!